sequential-agent/requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-cloud-storage>=3.0.0
google-cloud-secret-manager>=2.16.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
pyyaml>=6.0
requests>=2.31.0
